requirements.txt
# Core dependencies
pyserial>=3.5
pillow>=10.0.0
rich>=13.0.0
typer[all]>=0.9.0

# Optional: Streamlit web UI
streamlit>=1.28.0

# Development
pytest>=7.0
pytest-cov>=4.0
black>=23.0
ruff>=0.1.0
